Can create a tuple
